+++ src/DeviceManager.ts
@@ -5,17 +5,26 @@ export default class DeviceManager{
     private _isDebug:   boolean;
 
     // Elements
-    private _html:  HTMLElement;
-    private _body:  HTMLElement;
+    private _html:HTMLElement;
+    private _body:HTMLElement;
 
     // Custom Touch Input
-    private _trackedElements:   Array<Element>;
+    private _trackedElements:Array<Element>;
+
+    // Device/Connection Information
+    private _navigator:Navigator;
+    public static connection:Manager.NetworkInformation = undefined;
 
     constructor(debug?:boolean, setStatusClasses?:boolean){
         this._isDebug   = (debug) ? debug : false;
 
         this._html  = document.documentElement;
         this._body  = document.body;
+
+        this._navigator = window.navigator;
+        // @ts-ignore
+        DeviceManager.connection = this._navigator.connection || this._navigator.mozConnection || this._navigator.webkitConnection;
+        DeviceManager.connection.addEventListener('change', this.handleConnectionChange);
 
         if(setStatusClasses){
             this.setStatusClasses();
@@ -48,6 +57,16 @@ export default class DeviceManager{
     private userReleasedTouchedElement: EventListener = (e:Event)=>{
         const target = <HTMLElement>e.currentTarget;
         target.setAttribute('data-touching', "false");
+    }
+
+    /**
+     * Called when the `change` event is fired on `NetworkInformation`.
+     * @see https://developer.mozilla.org/en-US/docs/Web/API/NetworkInformation#Event_handlers
+     */
+    private handleConnectionChange:EventListener = (e:Event)=>{
+        this._navigator = window.navigator;
+        // @ts-ignore
+        DeviceManager.connection = this._navigator.connection || this._navigator.mozConnection || this._navigator.webkitConnection;
     }
 
     /**
@@ -227,6 +246,15 @@ export default class DeviceManager{
                 console.log('%c[Device Manager] '+`%cOpera: %c${ DeviceManager.isOpera }`,'color:#35ffb8','color:#eee', 'color:#68e5ff');
             }
         }
+
+        // Sets a status class if the device's connection type is known
+        if(DeviceManager.connection !== undefined){
+            this._html.classList.add(`is-${ DeviceManager.connection.effectiveType }`);
+
+            if(this._isDebug){
+                console.log('%c[Device Manager] '+`%cConnection Type: %c${ DeviceManager.connection.effectiveType }`,'color:#35ffb8','color:#eee', 'color:#68e5ff');
+            }
+        }
     }
     
     /**
@@ -275,7 +303,7 @@ export default class DeviceManager{
      * Checks if the browser is Edge 20+.
      * @returns `boolean`
      */
-    private static isEdge =(()=>{
+    public static isEdge =(()=>{
         let isEdge = false;
 
         // @ts-ignore
@@ -290,7 +318,7 @@ export default class DeviceManager{
      * Checks if the browser is Internet Explorer 6 - 11.
      * @returns `boolean`
      */
-    private static isIE = (()=>{
+    public static isIE = (()=>{
         let isIE = false;
 
         // @ts-ignore
@@ -305,7 +333,7 @@ export default class DeviceManager{
      * Checks if the browser is Firefox 1+.
      * @returns `boolean`
      */
-    private static isFirefox = (()=>{
+    public static isFirefox = (()=>{
         let isFirefox = false;
 
         // @ts-ignore
@@ -320,7 +348,7 @@ export default class DeviceManager{
      * Checks if the browser is Safari 3+.
      * @returns `boolean`
      */
-    private static isSafari = (()=>{
+    public static isSafari = (()=>{
         let isSafari = false;
 
         // @ts-ignore
@@ -335,7 +363,7 @@ export default class DeviceManager{
      * Checks if the browser is Opera 8+.
      * @returns `boolean`
      */
-    private static isOpera = (()=>{
+    public static isOpera = (()=>{
         let isOpera = false;
 
         // @ts-ignore
@@ -351,7 +379,7 @@ export default class DeviceManager{
      * @see https://en.wikipedia.org/wiki/Blink_(browser_engine)
      * @returns `boolean`
      */
-    private static isBlinkEngine = (()=>{
+    public static isBlinkEngine = (()=>{
         let isBlink = false;
 
         // @ts-ignore
@@ -366,7 +394,7 @@ export default class DeviceManager{
      * Checks if the browser supports touch input.
      * @returns `boolean`
      */
-    private static supportsTouch = (()=>{
+    public static supportsTouch = (()=>{
         let isTouchSupported = false;
 
         if(('ontouchstart' in window) || (navigator.maxTouchPoints > 0) || (navigator.msMaxTouchPoints > 0)){